# swagger.yaml
openapi: 3.0.3
info:
  title: Ashley AI Health Assistant API
  description: |
    A comprehensive AI-powered health assistant API that provides personalized health insights by combining user health data with medical knowledge.
    
    ## Features
    - Personalized health analysis based on user metrics
    - Medical knowledge retrieval and citation
    - Real-time streaming responses
    - Safety checks and anomaly detection
    - Conversation memory and context management
    
    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header.
    
    ## Rate Limiting
    API calls are rate limited to prevent abuse. Contact support for higher limits.
    
    ## Data Privacy
    All user data is encrypted and stored securely. No data is shared between users.
  version: 1.0.0
  contact:
    name: Ashley API Support
    email: support@ashley-health.com
    url: https://ashley-health.com/support
  license:
    name: Proprietary
    url: https://ashley-health.com/license

servers:
  - url: http://localhost:8088/v1
    description: Local development server
  - url: https://api.ashley-health.com/v1
    description: Production server
  - url: https://staging-api.ashley-health.com/v1
    description: Staging server

tags:
  - name: Health
    description: Health analysis and chat endpoints
  - name: Knowledge
    description: Knowledge base management
  - name: Configuration
    description: Product manager configuration endpoints
  - name: System
    description: System health and status

paths:
  /healthz:
    get:
      tags:
        - System
      summary: Health Check
      description: Basic health check endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
              example:
                status: "ok"

  /ready:
    get:
      tags:
        - System
      summary: Readiness Check
      description: Detailed readiness check for all dependencies
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ready"
              example:
                status: "ready"
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "not ready"
                  errors:
                    type: array
                    items:
                      type: string

  /chat:
    post:
      tags:
        - Health
      summary: Chat with Health Assistant
      description: |
        Send a health-related question and receive a personalized response based on your health data and medical knowledge.
        
        The response includes:
        - Personalized analysis based on your health metrics
        - Relevant medical information with citations
        - Safety warnings if applicable
        - Statistical analysis of your data
      operationId: chat
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              basic_question:
                summary: Basic health question
                value:
                  user_id: "user123"
                  message: "Why is my heart rate high today?"
              detailed_analysis:
                summary: Detailed analysis with timeframe
                value:
                  user_id: "user123"
                  message: "Analyze my sleep patterns over the last week"
                  timeframe:
                    start: "2024-01-01T00:00:00Z"
                    end: "2024-01-08T00:00:00Z"
                  metric_kinds: ["sleep", "hr", "hrv"]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                successful_response:
                  summary: Successful health analysis
                  value:
                    reply: "Based on your recent heart rate data, I notice your average HR was 85 BPM over the last 7 days, which is slightly elevated. This could be due to several factors including stress, caffeine intake, or lack of sleep. [1] Research shows that consistent elevated heart rate may indicate the need for lifestyle adjustments. [2] I recommend monitoring your stress levels and ensuring adequate sleep.\n\n**Your recent metrics:**\n- HR: mean 85.2, σ 12.1 over 42 points\n- SLEEP: mean 6.8 hours over 7 days\n\n**Sources:**\n[1] American Heart Association Guidelines (sim 0.892)\n[2] Sleep Foundation Research (sim 0.856)\n\n*Note: This is general information based on your data and referenced materials, not medical advice. For diagnosis, dosing, or urgent issues, consult a clinician or seek in-person care.*"
                    used_docs:
                      - id: "doc_001"
                        source: "American Heart Association Guidelines"
                        score: 0.892
                      - id: "doc_002"
                        source: "Sleep Foundation Research"
                        score: 0.856
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chat/stream:
    post:
      tags:
        - Health
      summary: Stream Chat Response
      description: |
        Send a health-related question and receive a real-time streaming response.
        
        The response is streamed as Server-Sent Events (SSE) with the following event types:
        - `node`: Progress updates as the AI processes your question
        - `final`: The complete answer
        - `error`: Error information if something goes wrong
        - `[DONE]`: End of stream marker
      operationId: chatStream
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Streaming response
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              examples:
                streaming_response:
                  summary: Example SSE stream
                  value: |
                    data: {"type": "node", "node": "parse_user", "keys": ["question", "metric_kinds"]}
                    
                    data: {"type": "node", "node": "pull_metrics", "keys": ["metrics"]}
                    
                    data: {"type": "node", "node": "analyze_metrics", "keys": ["stats", "anomalies"]}
                    
                    data: {"type": "node", "node": "retrieve_guidance", "keys": ["relevant_chunks", "citations"]}
                    
                    data: {"type": "node", "node": "safety", "keys": ["safety_warnings"]}
                    
                    data: {"type": "node", "node": "answer", "keys": ["answer"]}
                    
                    data: {"type": "final", "answer": "Based on your recent heart rate data..."}
                    
                    data: [DONE]
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /index/upsert:
    post:
      tags:
        - Knowledge
      summary: Add Knowledge Documents
      description: |
        Add or update documents in the knowledge base. This is typically used by administrators to populate the medical knowledge base.
        
        Documents are automatically embedded and indexed for semantic search.
      operationId: upsertDocuments
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRequest'
            examples:
              medical_document:
                summary: Add medical document
                value:
                  items:
                    - text: "Heart rate variability (HRV) is a measure of the variation in time between heartbeats. Higher HRV is generally associated with better cardiovascular health and fitness."
                      metadata:
                        source: "American Heart Association"
                        title: "Heart Rate Variability Guidelines"
                        category: "cardiovascular"
                        url: "https://www.heart.org/en/health-topics/high-blood-pressure"
                      id: "hrv_guidelines_001"
      responses:
        '200':
          description: Documents successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  ids:
                    type: array
                    items:
                      type: string
                    example: ["hrv_guidelines_001", "auto_generated_id_002"]
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /index/delete:
    post:
      tags:
        - Knowledge
      summary: Delete Knowledge Documents
      description: |
        Remove documents from the knowledge base by their IDs.
      operationId: deleteDocuments
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
            examples:
              delete_documents:
                summary: Delete specific documents
                value:
                  ids: ["hrv_guidelines_001", "outdated_doc_002"]
      responses:
        '200':
          description: Documents successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /index/search:
    post:
      tags:
        - Knowledge
      summary: Search Knowledge Base
      description: |
        Search the knowledge base for relevant documents using semantic search.
        
        This endpoint is useful for:
        - Finding relevant medical information
        - Testing knowledge base content
        - Building custom search interfaces
      operationId: searchKnowledge
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              search_heart_rate:
                summary: Search for heart rate information
                value:
                  query: "What is a normal heart rate range for adults?"
                  k: 5
              search_with_filter:
                summary: Search with metadata filter
                value:
                  query: "sleep recommendations"
                  k: 3
                  where:
                    key: "category"
                    value: "sleep"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                search_results:
                  summary: Example search results
                  value:
                    results:
                      - text: "Normal resting heart rate for adults ranges from 60 to 100 beats per minute (bpm). Athletes may have resting heart rates as low as 40 bpm."
                        metadata:
                          source: "American Heart Association"
                          title: "Understanding Heart Rate"
                          category: "cardiovascular"
                          score: 0.95
                      - text: "Heart rate variability (HRV) measures the variation in time between heartbeats and is an indicator of cardiovascular health."
                        metadata:
                          source: "Mayo Clinic"
                          title: "Heart Rate Variability"
                          category: "cardiovascular"
                          score: 0.87
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /config/metrics:
    get:
      tags:
        - Configuration
      summary: Get Metric Configuration
      description: |
        Retrieve the current metric configuration including default and available metric kinds.
        
        This endpoint is useful for product managers to understand what metrics
        are currently configured for health analysis.
      operationId: getMetricConfig
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Current metric configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricConfigResponse'
              examples:
                current_config:
                  summary: Current metric configuration
                  value:
                    default_metric_kinds: ["hr", "hrv", "steps", "sleep"]
                    available_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature", "glucose", "oxygen_saturation"]
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    post:
      tags:
        - Configuration
      summary: Update Metric Configuration
      description: |
        Update the metric configuration for health analysis.
        
        Allows product managers to adjust which metrics are analyzed by default
        and which metrics are available for analysis.
        
        **Important Notes:**
        - Changes require application restart to take effect in production
        - For production deployments, consider using environment variables
        - All default metric kinds must be included in available metric kinds
        - Common metric kinds: hr, hrv, steps, sleep, weight, blood_pressure, temperature, glucose, oxygen_saturation
      operationId: updateMetricConfig
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetricConfigUpdateRequest'
            examples:
              update_defaults:
                summary: Update default metric kinds
                value:
                  default_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight"]
              add_new_metrics:
                summary: Add new available metrics
                value:
                  available_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature", "glucose", "oxygen_saturation", "blood_glucose", "cholesterol"]
              full_update:
                summary: Update both default and available metrics
                value:
                  default_metric_kinds: ["hr", "hrv", "steps", "sleep"]
                  available_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature"]
      responses:
        '200':
          description: Metric configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricConfigResponse'
              examples:
                updated_config:
                  summary: Updated configuration
                  value:
                    default_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight"]
                    available_metric_kinds: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature", "glucose", "oxygen_saturation"]
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Invalid metric configuration
                  value:
                    error: "Default metric kinds ['invalid_metric'] are not in available metric kinds"
                    details: "All default metric kinds must be included in the available metric kinds list"
                    code: "VALIDATION_ERROR"
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    Timeframe:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
          description: ISO8601 UTC start time
          example: "2024-01-01T00:00:00Z"
        end:
          type: string
          format: date-time
          description: ISO8601 UTC end time
          example: "2024-01-08T00:00:00Z"

    ChatRequest:
      type: object
      required:
        - user_id
        - message
      properties:
        user_id:
          type: string
          description: Unique identifier for the user
          example: "user123"
        message:
          type: string
          description: The health question or message
          example: "Why is my heart rate high today?"
          maxLength: 2000
        metadata:
          type: object
          description: Additional metadata for the request
          additionalProperties: true
        timeframe:
          $ref: '#/components/schemas/Timeframe'
        metric_kinds:
          type: array
          items:
            type: string
          description: Types of health metrics to analyze
          example: ["hr", "hrv", "steps", "sleep"]
          enum: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature"]

    ChatResponse:
      type: object
      required:
        - reply
      properties:
        reply:
          type: string
          description: The AI assistant's response
          example: "Based on your recent heart rate data, I notice your average HR was 85 BPM over the last 7 days..."
        used_docs:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
          description: Sources and citations used in the response

    Citation:
      type: object
      properties:
        id:
          type: string
          description: Document ID
          example: "doc_001"
        source:
          type: string
          description: Source name or URL
          example: "American Heart Association Guidelines"
        score:
          type: number
          format: float
          description: Relevance score (0-1)
          example: 0.892

    UpsertItem:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Document content
          example: "Heart rate variability (HRV) is a measure of the variation in time between heartbeats..."
        metadata:
          type: object
          description: Document metadata
          additionalProperties: true
          example:
            source: "American Heart Association"
            title: "Heart Rate Variability Guidelines"
            category: "cardiovascular"
        id:
          type: string
          description: Optional document ID (auto-generated if not provided)
          example: "hrv_guidelines_001"

    UpsertRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UpsertItem'
          description: List of documents to add or update

    DeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: string
          description: List of document IDs to delete
          example: ["doc_001", "doc_002"]

    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query
          example: "What is a normal heart rate range?"
        k:
          type: integer
          description: Number of results to return
          default: 5
          minimum: 1
          maximum: 50
        where:
          type: object
          description: Metadata filter
          properties:
            key:
              type: string
              example: "category"
            value:
              type: string
              example: "cardiovascular"

    SearchResultItem:
      type: object
      required:
        - text
        - metadata
      properties:
        text:
          type: string
          description: Document content
          example: "Normal resting heart rate for adults ranges from 60 to 100 beats per minute..."
        metadata:
          type: object
          description: Document metadata including relevance score
          additionalProperties: true
          example:
            source: "American Heart Association"
            title: "Understanding Heart Rate"
            category: "cardiovascular"
            score: 0.95

    SearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
          description: Search results

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
          example: "The 'user_id' field is required"
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"

    MetricConfigResponse:
      type: object
      required:
        - default_metric_kinds
        - available_metric_kinds
      properties:
        default_metric_kinds:
          type: array
          items:
            type: string
          description: Default metric kinds analyzed when not specified in requests
          example: ["hr", "hrv", "steps", "sleep"]
        available_metric_kinds:
          type: array
          items:
            type: string
          description: All available metric kinds that can be analyzed
          example: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature", "glucose", "oxygen_saturation"]

    MetricConfigUpdateRequest:
      type: object
      properties:
        default_metric_kinds:
          type: array
          items:
            type: string
          description: New default metric kinds (optional)
          example: ["hr", "hrv", "steps", "sleep", "weight"]
        available_metric_kinds:
          type: array
          items:
            type: string
          description: New available metric kinds (optional)
          example: ["hr", "hrv", "steps", "sleep", "weight", "blood_pressure", "temperature", "glucose", "oxygen_saturation"]

  examples:
    HealthQuestion:
      summary: Typical health question
      value:
        user_id: "user123"
        message: "Why is my heart rate high today?"
        metric_kinds: ["hr", "hrv"]
        timeframe:
          start: "2024-01-01T00:00:00Z"
          end: "2024-01-08T00:00:00Z"

    MedicalDocument:
      summary: Medical knowledge document
      value:
        text: "Heart rate variability (HRV) is a measure of the variation in time between heartbeats. Higher HRV is generally associated with better cardiovascular health and fitness."
        metadata:
          source: "American Heart Association"
          title: "Heart Rate Variability Guidelines"
          category: "cardiovascular"
          url: "https://www.heart.org/en/health-topics/high-blood-pressure"
        id: "hrv_guidelines_001"